extract BaseProxy to deduplicate filtering logic and fix prompt convention
